Move arch prefix helper to its own package
