+++ pkg/arch/arch_test.go
@@ -1,4 +1,4 @@
-package tracing
+package arch
 
 import (
 	"testing"
